cleaning: move helpers file and remove unused files
