+++ v-next/hardhat-chai-matchers/test/fixture-projects/hardhat-project/hardhat.config.ts
@@ -1,0 +1,5 @@
+import type { HardhatUserConfig } from "@ignored/hardhat-vnext/config";
+
+const config: HardhatUserConfig = {};
+
+export default config;
