--- v-next/hardhat-chai-matchers/test/fixture-projects/hardhat-project/start-node.js
@@ -1,14 +1,0 @@
-const hre = require("hardhat");
-
-async function main() {
-  await hre.run("node", {
-    port: +process.env.HARDHAT_NODE_PORT,
-  });
-}
-
-main()
-  .then(() => process.exit(0))
-  .catch((error) => {
-    console.error(error);
-    process.exit(1);
-  });
