+++ v-next/hardhat-chai-matchers/test/reverted/revertedWithPanic.ts
@@ -21,7 +21,7 @@ import {
   runSuccessfulAsserts,
   runFailedAsserts,
   mineSuccessfulTransaction,
-} from "../helpers.js";
+} from "../helpers/helpers.js";
 
 addChaiMatchers();
